GameManager

+++ Assets/Scripts/Brick Spawner/BrickMove.cs
@@ -6,20 +6,21 @@ public class BrickMove : MonoBehaviour
 {
     //Drag Player Transform Here
     public Transform target;
-    public int speed = 3;
+    public float speed = 3;
+
     void Start()
     {
-       MoveBrick();
+        MoveBrick();
     }
 
     // Update is called once per frame
     void Update()
     {
-        
+
     }
 
     public void MoveBrick()
     {
-        transform.position = Vector3.MoveTowards(transform.position,target.position, speed * Time.deltaTime);
+        transform.position = Vector3.MoveTowards(transform.position, target.position, speed * Time.deltaTime);
     }
 }
